refactor(cms-models):将 isTeacher 字段类型从 Long 改为 Integer
- 修改了 ApplyForTeacher 类中的 isTeacher 字段类型
- 优化了字段类型，使其更符合实际需求

+++ cms-models/cms-models-xiangmu/src/main/java/com/xk/domain/dto/ApplyForTeacher.java
@@ -15,7 +15,7 @@ public class ApplyForTeacher {
     @NotNull(message = "请选择是否指导老师")
     @Min(value = 0, message = "isTeacher输入0或者1代表是否指导老师或者企业老师")
     @Max(value = 1, message = "isTeacher输入0或者1代表是否指导老师或者企业老师")
-    private Long isTeacher;
+    private Integer isTeacher;
     /**
      * 邮箱
      */
